backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
httpx==0.26.0
google-generativeai==0.3.2
websockets==12.0
pillow==10.2.0
osmnx==1.8.1
networkx==3.2.1
geopandas==0.14.2
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
# MATLAB Engine API - Install separately from MATLAB installation
# See: https://www.mathworks.com/help/matlab/matlab_external/install-the-matlab-engine-for-python.html
